In progress: Call local backend api to make get Request to Steam api.
